# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sampler Book Library API",
        "description": "This is the documentation of the Sampler Book Library API, with this API you can check-in and out books after you register as an user",
        "version": "0.0.1"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Register as new user.",
                "description": "Register as new user",
                "operationId": "App\\Http\\Controllers\\AuthController::register",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Register as new user.",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "Register as new user.",
                                        "type": "string"
                                    },
                                    "date_of_birth": {
                                        "description": "Register as new user.",
                                        "type": "date"
                                    },
                                    "password": {
                                        "description": "Register as new user.",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "George Lucas",
                                    "email": "george@lucasarts.com",
                                    "date_of_birth": "1944-05-14",
                                    "password": "5T4r W4Rs"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An acess_token response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "JWT access token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Token type",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "Token expiration in seconds",
                                            "type": "integer",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGc...",
                                        "token_type": "bearer",
                                        "expires_in": 3600
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Register as new user.",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Register as new user.",
                                            "properties": {
                                                "email": {
                                                    "description": "Register as new user.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email must be a valid email address."
                                                    }
                                                },
                                                "date_of_birth": {
                                                    "description": "Register as new user.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The date of birth is not a valid date."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Get a JWT via given credentials.",
                "description": "Endpoint for log in",
                "operationId": "App\\Http\\Controllers\\AuthController::login",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "description": "Get a JWT via given credentials.",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "Get a JWT via given credentials.",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "email": "foo@foobar.com",
                                    "password": "P455w0Rd"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "An acess_token response.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "description": "JWT access token",
                                            "type": "string"
                                        },
                                        "token_type": {
                                            "description": "Token type",
                                            "type": "string"
                                        },
                                        "expires_in": {
                                            "description": "Token expiration in seconds",
                                            "type": "integer",
                                            "items": {}
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "access_token": "eyJ0eXAiOiJKV1QiLCJhbGc...",
                                        "token_type": "bearer",
                                        "expires_in": 3600
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Get a JWT via given credentials.",
                                            "type": "string",
                                            "example": "The given data was invalid."
                                        },
                                        "errors": {
                                            "description": "Get a JWT via given credentials.",
                                            "properties": {
                                                "email": {
                                                    "description": "Get a JWT via given credentials.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email field is required."
                                                    }
                                                },
                                                "password": {
                                                    "description": "Get a JWT via given credentials.",
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The password field is required."
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Log the user out (Invalidate the token).",
                "description": "Endpoint for log out. It invalidates the JWT",
                "operationId": "App\\Http\\Controllers\\AuthController::logout",
                "responses": {
                    "200": {
                        "description": "Successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "JWT access token",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object",
                                    "example": {
                                        "message": "Successfully logged out"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "access_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "access_token": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}